AutoTest拆解進度 60% Everything looks fine!  Yeah!!
For now.
