+++ HTMLTestRunner.py
@@ -670,7 +670,7 @@ class HTMLTestRunner(Template_mixin):
         test(result)
         self.stopTime = datetime.datetime.now()
         self.generateReport(test, result)
-        #print(sys.stderr, '\nTime Elapsed: %s' % (self.stopTime-self.startTime))
+        #print(sys.stderr, '\nTime Elapsed: %s' % (srows.append(row)elf.stopTime-self.startTime))
         return result
